fish: fix sourcing order
As files under conf.d are sourced before config.fish, PATH isn't set
the first time fish sources `aliases.fish`, so there are no external
(brew) commands available. It is then necessary to run `exec fish` in
order to have working aliases for the session, as env variables are
preserved.

As this is inconvenient, define a function on `alises.fish` to be
sourced after PATH has been properly set.